% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceltoolsV3.R
\name{createexcelfile}
\alias{createexcelfile}
\title{Create an excel file and save your tables as sheets in that document}
\usage{
createexcelfile(title, ..., overwrite = FALSE)
}
\arguments{
\item{title}{A filename for the excel file to be created}

\item{...}{A number of dataframes to be used as sheets for the new excel file}

\item{overwrite}{A logical toggle to choose whether to overwrite a potential existing file}
}
\value{
A tibble containing the data from the chosen sheet
}
\description{
As of yet there is no way of giving the sheets other names.
The sheets in the resulting document will therefore inherit the variable names given in this function.
e.g. table1 will give rise to a sheet in the excel document called table1.
}
