% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceltoolsV3.R
\name{createexcelfile}
\alias{createexcelfile}
\title{load specific sheets from a chosen excell file}
\usage{
createexcelfile(title, ..., overwrite = TRUE)
}
\arguments{
\item{title}{A filename for the excel file to be created}

\item{...}{A number of dataframes to be used as sheets for the new excel file}

\item{overwrite}{A logical toggle to choose whether to overwrite a potential existing file}
}
\value{
A tibble containing the data from the chosen sheet
}
\description{
Let you chose a file from a popup window, then lists the sheets to choose from
}
