% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceltoolsV3.R
\name{readsheet}
\alias{readsheet}
\title{load specific sheets from a chosen excel file}
\usage{
readsheet(filepath = NULL, sheet = NULL)
}
\arguments{
\item{filepath}{A path to an excel file}

\item{sheet}{An integer referring to a specific sheet to open}
}
\value{
A tibble containing the data from the chosen sheet
}
\description{
Let you chose a file from a popup window, then lists the sheets to choose from
}
